backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
python-dotenv==1.0.1
sqlparse==0.5.3
typing-extensions==4.13.1
